requirements.txt
customtkinter==5.2.0
Pillow==10.0.1
mysql-connector-python==8.1.0
cohere==4.34